pyproject.toml
[project]
name = "rec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.1",
    "sentence-transformers>=5.0.0",
    "standard-imghdr>=3.13.0",
    "tbparse>=0.0.9",
    "tensorboard>=2.19.0",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
]
